+++ subot/ui_areas/ui_ocr_types.py
@@ -1,5 +1,6 @@
 from typing import Union
 
+from subot.ui_areas.AnointmentClaimUI import AnointmentClaimUI
 from subot.ui_areas.CodexGeneric import CodexGeneric
 from subot.ui_areas.CreatureReorderSelectFirst import OCRCreatureRecorderSelectFirst, OCRCreatureRecorderSwapWith
 from subot.ui_areas.OCRGodForgeSelect import OCRGodForgeSelectSystem
@@ -10,6 +11,6 @@ from subot.ui_areas.summoning import OcrSummoningSystem
 from subot.ui_areas.OcrUnknownArea import OcrUnknownArea
 
 OCR_UI_SYSTEMS = Union[
-    OCRCreatureRecorderSelectFirst, OCRCreatureRecorderSwapWith, OCRGodForgeSelectSystem, OCRCreaturesDisplaySystem, OcrSummoningSystem, OcrUnknownArea, CodexGeneric, OCRRealmSelect, PerkScreen]
+    OCRCreatureRecorderSelectFirst, OCRCreatureRecorderSwapWith, OCRGodForgeSelectSystem, OCRCreaturesDisplaySystem, OcrSummoningSystem, OcrUnknownArea, CodexGeneric, OCRRealmSelect, PerkScreen, AnointmentClaimUI]